Split mpConfig-2.0 FATs which throw deployment exceptions
