+++ dev/io.openliberty.microprofile.config.2.0.internal_fat/fat/src/io/openliberty/microprofile/config/internal_fat/Config20Tests.java
@@ -10,19 +10,13 @@
  *******************************************************************************/
 package io.openliberty.microprofile.config.internal_fat;
 
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertTrue;
-
 import java.io.File;
-import java.util.List;
 
-import org.eclipse.microprofile.config.spi.Converter;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
-import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import com.ibm.websphere.simplicity.PropertiesAsset;
@@ -38,7 +32,6 @@ import componenttest.rules.repeater.RepeatTests;
 import componenttest.topology.impl.LibertyServer;
 import componenttest.topology.utils.FATServletClient;
 import io.openliberty.microprofile.config.internal_fat.apps.TestUtils;
-import io.openliberty.microprofile.config.internal_fat.apps.brokenInjection.ValidConverter;
 import io.openliberty.microprofile.config.internal_fat.apps.classLoader.ClassLoadersTestServlet;
 import io.openliberty.microprofile.config.internal_fat.apps.converter.ConvertersTestServlet;
 import io.openliberty.microprofile.config.internal_fat.apps.defaultSources.DefaultSourcesTestServlet;
@@ -47,9 +40,7 @@ import io.openliberty.microprofile.config.internal_fat.apps.propertyExpression.P
 @RunWith(FATRunner.class)
 public class Config20Tests extends FATServletClient {
 
-    public static final String BAD_OBSERVER_APP_NAME = "badObserverApp";
     public static final String DEFAULT_SOURCES_APP_NAME = "defaultSourcesApp";
-    public static final String BROKEN_INJECTION_APP_NAME = "brokenInjectionApp";
     public static final String CLASS_LOADER_APP_NAME = "classLoadersApp";
     public static final String CONVERTER_LOADER_APP_NAME = "convertersApp";
     public static final String PROPERTY_EXPRESSION_APP_NAME = "propertyExpressionApp";
@@ -71,13 +62,6 @@ public class Config20Tests extends FATServletClient {
 
     @BeforeClass
     public static void setUp() throws Exception {
-
-        WebArchive badObserverWar = ShrinkWrap.create(WebArchive.class, BAD_OBSERVER_APP_NAME + ".war")
-                        .addPackages(true, "io.openliberty.microprofile.config.internal_fat.apps.badobserver");
-
-        WebArchive brokenInjectionWar = ShrinkWrap.create(WebArchive.class, BROKEN_INJECTION_APP_NAME + ".war")
-                        .addPackages(true, "io.openliberty.microprofile.config.internal_fat.apps.brokenInjection")
-                        .addAsServiceProvider(Converter.class, ValidConverter.class);
 
         WebArchive defaultSourcesWar = ShrinkWrap.create(WebArchive.class, DEFAULT_SOURCES_APP_NAME + ".war")
                         .addPackages(true, DefaultSourcesTestServlet.class.getPackage())
@@ -105,10 +89,6 @@ public class Config20Tests extends FATServletClient {
                         .addPackages(true, PropertyExpressionTestServlet.class.getPackage())
                         .addAsResource(config, "META-INF/microprofile-config.properties");
 
-        // The first 2 wars should throw deployment exceptions, hence don't validate.
-        ShrinkHelper.exportDropinAppToServer(server, badObserverWar, DeployOptions.SERVER_ONLY, DeployOptions.DISABLE_VALIDATION);
-        ShrinkHelper.exportDropinAppToServer(server, brokenInjectionWar, DeployOptions.SERVER_ONLY, DeployOptions.DISABLE_VALIDATION);
-
         ShrinkHelper.exportDropinAppToServer(server, defaultSourcesWar, DeployOptions.SERVER_ONLY);
         ShrinkHelper.exportDropinAppToServer(server, classLoadersWar, DeployOptions.SERVER_ONLY);
         ShrinkHelper.exportDropinAppToServer(server, convertersWar, DeployOptions.SERVER_ONLY);
@@ -116,54 +96,6 @@ public class Config20Tests extends FATServletClient {
 
         server.startServer();
 
-    }
-
-    @Test
-    public void testBadObserver() throws Exception {
-        List<String> errors = server
-                        .findStringsInLogs("SRCFG02000: No Config Value exists for required property DOESNOTEXIST");
-        assertNotNull("error not found", errors);
-        assertTrue("error not found: " + errors.size(), errors.size() > 0);
-    }
-
-    @Test
-    public void testMethodUnnamed() throws Exception {
-        List<String> errors = server
-                        .findStringsInLogs("SRCFG02002: Could not find default name for .*io.openliberty.microprofile.config.internal_fat.apps.brokenInjection.ConfigUnnamedMethodInjectionBean.*setSimpleKey6");
-        assertNotNull("error not found", errors);
-        assertTrue("error not found: " + errors.size(), errors.size() > 0);
-    }
-
-    @Test
-    public void testConstructorUnnamed() throws Exception {
-        List<String> errors = server
-                        .findStringsInLogs("SRCFG02002: Could not find default name for .*io.openliberty.microprofile.config.internal_fat.apps.brokenInjection.ConfigUnnamedConstructorInjectionBean");
-        assertNotNull("error not found", errors);
-        assertTrue("error not found: " + errors.size(), errors.size() > 0);
-    }
-
-    @Test
-    public void testNonExistantKey() throws Exception {
-        List<String> errors = server
-                        .findStringsInLogs("SRCFG02000: No Config Value exists for required property io.openliberty.microprofile.config.internal_fat.apps.brokenInjection.MissingConfigPropertyBean.nonExistantKey");
-        assertNotNull(errors);
-        assertTrue(errors.size() > 0);
-    }
-
-    @Test
-    public void testNonExistantKeyWithCustomConverter() throws Exception {
-        List<String> errors = server
-                        .findStringsInLogs("SRCFG02000: No Config Value exists for required property io.openliberty.microprofile.config.internal_fat.apps.brokenInjection.MissingConfigPropertyBean.undefinedKeyWithConverter");
-        assertNotNull(errors);
-        assertTrue(errors.size() > 0);
-    }
-
-    @Test
-    public void testConverterMissing() throws Exception {
-        List<String> errors = server
-                        .findStringsInLogs("SRCFG02006: The property noConverterKey cannot be converted to class io.openliberty.microprofile.config.internal_fat.apps.brokenInjection.TypeWithNoConverter");
-        assertNotNull(errors);
-        assertTrue(errors.size() > 0);
     }
 
     @AfterClass
