+++ dev/io.openliberty.microprofile.config.2.0.internal_fat/publish/servers/Config20ExceptionServer/bootstrap.properties
@@ -1,0 +1,6 @@
+bootstrap.include=../testports.properties
+osgi.console=7773
+org.jboss.logging.provider=slf4j
+io.openliberty.microprofile.config.internal_fat.apps.brokenInjection.ConfigUnnamedConstructorInjectionBean.SIMPLE_KEY4=VALUE4
+io.openliberty.microprofile.config.internal_fat.apps.brokenInjection.ConfigUnnamedMethodInjectionBean.SIMPLE_KEY6=VALUE6
+noConverterKey=value
