+++ dev/io.openliberty.microprofile.config.2.0.internal_fat/publish/servers/Config20Server/bootstrap.properties
@@ -1,10 +1,7 @@
 bootstrap.include=../testports.properties
 osgi.console=7773
-io.openliberty.microprofile.config.internal_fat.apps.brokenInjection.ConfigUnnamedConstructorInjectionBean.SIMPLE_KEY4=VALUE4
-io.openliberty.microprofile.config.internal_fat.apps.brokenInjection.ConfigUnnamedMethodInjectionBean.SIMPLE_KEY6=VALUE6
+org.jboss.logging.provider=slf4j
 DUPLICATE_CONVERTERS_KEY_1=INPUT1
 DUPLICATE_CONVERTERS_KEY_2=INPUT2
 DUPLICATE_CONVERTERS_KEY_3=INPUT3
-noConverterKey=value
-org.jboss.logging.provider=slf4j
 value1DefinedInTwoPlaces=value1a
